Relocate errors file into errors folder
