+++ packages/gbnf/javascript/src/utils/errors/errors.test.ts
